Correction in registerActivity tests

+++ app/src/androidTest/java/com/sdp/swiftwallet/presentation/signIn/RegisterActivityTest.java
@@ -5,8 +5,6 @@ import static androidx.test.espresso.action.ViewActions.click;
 import static androidx.test.espresso.action.ViewActions.closeSoftKeyboard;
 import static androidx.test.espresso.action.ViewActions.typeText;
 import static androidx.test.espresso.assertion.ViewAssertions.matches;
-import static androidx.test.espresso.intent.Intents.intended;
-import static androidx.test.espresso.intent.matcher.IntentMatchers.toPackage;
 import static androidx.test.espresso.matcher.ViewMatchers.hasFocus;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
@@ -137,9 +135,9 @@ public class RegisterActivityTest {
 
     @Test
     public void registerUserFiresLoginIntentCorrectly() {
-        String userTestUsername = "usernameTest";
-        String userTestEmail = "email.test@epfl.ch";
-        String userTestPassword = "Password1";
+        String userTestUsername = "registerTest";
+        String userTestEmail = "register.test@epfl.ch";
+        String userTestPassword = "Register1";
 
         deleteUser(userTestEmail, userTestPassword);
 
@@ -147,6 +145,7 @@ public class RegisterActivityTest {
         onView(withId(R.id.registerInputUsername)).perform(typeText(userTestUsername), closeSoftKeyboard());
         onView(withId(R.id.registerInputEmail)).perform(typeText(userTestEmail), closeSoftKeyboard());
         onView(withId(R.id.registerInputPassword)).perform(typeText(userTestPassword), closeSoftKeyboard());
+        onView(withId(R.id.registerInputConfirmPassword)).perform(typeText(userTestPassword), closeSoftKeyboard());
 
         onView(withId(R.id.registerBtn)).perform(click());
         //intended(toPackage("com.sdp.swiftwallet"));
@@ -154,14 +153,12 @@ public class RegisterActivityTest {
 
     // Used by registerUserFiresLoginIntentCorrectly test
     private void deleteUser(String userTestEmail, String userTestPassword) {
-        // SignIn user to test and delete it if it exists
         mIdlingResource.increment();
-        FirebaseAuth mAuth = FirebaseAuth.getInstance();
-        if (mAuth.getCurrentUser() == null) {
-            mAuth.signInWithEmailAndPassword(userTestEmail, userTestPassword)
+        if (db.getCurrentUser() == null) {
+            db.signInWithEmailAndPassword(userTestEmail, userTestPassword)
                     .addOnCompleteListener(task -> {
                         if (task.isSuccessful()) {
-                            FirebaseUser user = mAuth.getCurrentUser();
+                            FirebaseUser user = db.getCurrentUser();
                             user.delete().addOnCompleteListener(task1 ->
                                     mIdlingResource.decrement()
                             );
@@ -170,16 +167,18 @@ public class RegisterActivityTest {
                         }
                     });
         } else {
-            FirebaseUser user = mAuth.getCurrentUser();
-            user.delete().addOnCompleteListener(task -> mIdlingResource.decrement());
+            if (db.getCurrentUser().getEmail().equals(userTestEmail)) {
+                FirebaseUser user = db.getCurrentUser();
+                user.delete().addOnCompleteListener(task -> mIdlingResource.decrement());
+            }
         }
     }
 
     @Test
     public void registerUserFailsCorrectly() {
-        String userTestUsername = "usernameTest";
-        String userTestEmail = "email.test@epfl.ch";
-        String userTestPassword = "Password1";
+        String userTestUsername = "registerTest";
+        String userTestEmail = "register.test@epfl.ch";
+        String userTestPassword = "Register1";
 
         // Create user before creating it for test so that it fails
         createUser(userTestEmail, userTestPassword);
@@ -188,6 +187,7 @@ public class RegisterActivityTest {
         onView(withId(R.id.registerInputUsername)).perform(typeText(userTestUsername), closeSoftKeyboard());
         onView(withId(R.id.registerInputEmail)).perform(typeText(userTestEmail), closeSoftKeyboard());
         onView(withId(R.id.registerInputPassword)).perform(typeText(userTestPassword), closeSoftKeyboard());
+        onView(withId(R.id.registerInputConfirmPassword)).perform(typeText(userTestPassword), closeSoftKeyboard());
 
         onView(withId(R.id.registerBtn)).perform(click());
         onView(withId(R.id.registerBtn)).check(matches(isDisplayed()));
